requirements.txt
transformers>=4.30.0
torch>=2.0.0
torchaudio>=2.0.0
ffmpeg-python>=0.2.0
rich>=13.0.0
datasets>=2.13.0
accelerate>=0.21.0